refactor: ChattedDto package
