+++ src/main/java/com/gaaji/chat/statusmanagement/domain/controller/dto/ChattedDto.java
@@ -1,4 +1,4 @@
-package com.gaaji.chat.statusmanagement.domain.controller;
+package com.gaaji.chat.statusmanagement.domain.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
