+++ src/main/java/com/gaaji/chat/statusmanagement/domain/service/NotificationServiceImpl.java
@@ -1,7 +1,7 @@
 package com.gaaji.chat.statusmanagement.domain.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.gaaji.chat.statusmanagement.domain.controller.ChattedDto;
+import com.gaaji.chat.statusmanagement.domain.controller.dto.ChattedDto;
 import com.gaaji.chat.statusmanagement.domain.entity.ChatRoom;
 import com.gaaji.chat.statusmanagement.domain.entity.RoomId;
 import com.gaaji.chat.statusmanagement.domain.repository.ChatRoomRepository;
